Tower placement on cross, slow targeting.
Fixed towers being able to be placed at crossroads. Turns out it was me
being lazy and using type 0 for the path.

Fixed slows not being able to target when towers are constantly going
out of it's range. For some reason I reset the bullet wait timer every
time a new target was acquired. May cause other bugs, it's line 170 in
Tower if it needs to be uncommented (At time of writing).

Changed the corner graphics to fit with the other tiles better.